+++ internal/features/tasks/service/service.go
@@ -24,7 +24,7 @@ type TaskRepository interface {
 
 	GetTask(
 		ctx context.Context,
-		taskID int,
+		taskID int64,
 	) (core_domain.Task, error)
 
 	CompleteTask(
